+++ src/engine/Actions/Action/MoveBy.ts
@@ -62,5 +62,6 @@ export class MoveBy implements Action {
 
   public reset(): void {
     this._started = false;
+    this._stopped = false;
   }
 }