+++ src/engine/Actions/Action/ScaleBy.ts
@@ -51,8 +51,8 @@ export class ScaleBy implements Action {
   public isComplete(): boolean {
     return (
       this._stopped ||
-      (Math.abs(this._tx.scale.x - this._startScale.x) >= this._distanceX &&
-        Math.abs(this._tx.scale.y - this._startScale.y) >= this._distanceY)
+      (Math.abs(this._tx.scale.x - this._startScale.x) >= (this._distanceX - 0.01) &&
+        Math.abs(this._tx.scale.y - this._startScale.y) >= (this._distanceY - 0.01))
     );
   }
 
@@ -64,5 +64,6 @@ export class ScaleBy implements Action {
 
   public reset(): void {
     this._started = false;
+    this._stopped = false;
   }
 }